refactor(e2e): move testNotes to `pluginFixtures` and update imports
